Add optional onTap callback for Markers
This adds an optional onMarkerTap callback which gets triggered when
tapping a marker. My motivation is as follows:

I have an app which shows a popup when a marker is tapped. Originally I
implemented this by wrapping an Icon in a GestureBuilder in the Marker's
child Widget only to discover that only one GestureBuilder will get
triggered per gesture event (in this case tap). This meant that my
popups broke the centerMarkerOnClick functionality.