Use the marker's key to avoid rebuilding
The current behavior is that, when the marker list changes, all marker widgets are rebuilt. This is because the `MarkerWidget` uses an `ObjectKey` on the `MarkerNode`, which is always created anew.

The `MapWidget` is currently constructed with this a key as proposed in this change. I can't think of a reason why `MarkerWidget` should be different.